list_style/add first version of list
