+++ src/container/list/components/QuestionCard/CardContent/index.style.tsx
