+++ src/container/list/components/QuestionCard/CardHeader/index.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import { CardHeaderStyled } from './index.style';
+
+const CardHeader = ({ cardTitle }): React.ReactElement => {
+    return <CardHeaderStyled title={cardTitle} />;
+};
+
+export default CardHeader;
