+++ src/container/list/components/QuestionCard/index.style.tsx
