+++ src/container/list/components/QuestionList/index.style.tsx
@@ -1,0 +1,9 @@
+import styled from '@emotion/styled';
+import { Masonry } from '@mui/lab';
+
+export const QuestionListStyled = styled(Masonry)`
+    width: 700px;
+    background-color: rgba(16, 16, 16, 0.11);
+    border-radius: 30px;
+    padding-top: 1.5vh;
+`;
