--- src/container/list/components/masonry/masonry.style.tsx
@@ -1,6 +1,0 @@
-import styled from '@emotion/styled';
-import { Masonry } from '@mui/lab';
-
-export const MasonryStyled = styled(Masonry)`
-    width: 700px;
-`;
