--- src/container/list/components/masonry/masonry.tsx
@@ -1,20 +1,0 @@
-import QuestionCard from '../questionCard/questionCard';
-import React from 'react';
-import { MasonryBoxStyled } from './masonryBox.style';
-import { MasonryStyled } from './masonry.style';
-
-const heights = [400, 500, 600, 400, 400];
-
-const CardsMasonry = () => {
-    return (
-        <MasonryBoxStyled>
-            <MasonryStyled columns={2} spacing={2}>
-                {heights.map((height, index) => (
-                    <QuestionCard cardHeight={height} cardNum={index + 1}></QuestionCard>
-                ))}
-            </MasonryStyled>
-        </MasonryBoxStyled>
-    );
-};
-
-export default CardsMasonry;
