--- src/container/list/components/masonry/masonryBox.style.tsx
@@ -1,11 +1,0 @@
-import styled from '@emotion/styled';
-import { Box } from '@mui/material';
-
-export const MasonryBoxStyled = styled(Box)`
-    background-color: rgba(16, 16, 16, 0.11);
-    width: 100%;
-    border-radius: 30px;
-    padding-top: 1.5vh;
-    display: flex;
-    justify-content: center;
-`;
