+++ src/container/main/components/Main/index.style.tsx
@@ -3,4 +3,6 @@ import styled from '@emotion/styled';
 export const MainStyled = styled.main`
     // Какие-то стили для содержимого страницы (например шрифты)
     flex-grow: 1;
+    display: flex;
+    justify-content: center;
 `;
